requirements.txt
customtkinter
matplotlib
sympy
pandas
python-dotenv
google-genai
pyinstaller
PyQt6